修改：作者名字和插件命名
